refactored all user lists
